refactor: remove optimized screenshots route and integrate its functionality into the main screenshots route
