+++ apps/server/src/routes/screenshots/index.ts
@@ -1,5 +1,6 @@
 import { OpenAPIHono, createRoute, z } from "@hono/zod-openapi";
 import { CreateScreenshotSchema } from "@screenshothis/schemas/screenshots";
+import { endTime, setMetric, startTime } from "hono/timing";
 import { objectToCamel } from "ts-case-convert";
 
 import type { Variables } from "#/common/environment";
@@ -15,6 +16,7 @@ import {
 	enqueueScreenshotJob,
 	getExistingScreenshotKey,
 } from "#/lib/screenshot-queue";
+import { deduplicateRequest, generateCacheKey } from "#/utils/deduplication";
 import { createErrorResponse } from "#/utils/errors";
 
 const PLACEHOLDER_IMAGE = Buffer.from(
@@ -22,7 +24,9 @@ const PLACEHOLDER_IMAGE = Buffer.from(
 	"base64",
 );
 
-const screenshots = new OpenAPIHono<{ Variables: Variables }>().openapi(
+const optimizedScreenshots = new OpenAPIHono<{
+	Variables: Variables;
+}>().openapi(
 	createRoute({
 		method: "get",
 		path: "/take",
@@ -51,19 +55,24 @@ const screenshots = new OpenAPIHono<{ Variables: Variables }>().openapi(
 						},
 					},
 				},
-				description: "Successful response",
+				description: "Optimized screenshot response",
 			},
 		},
 	}),
 	async (c) => {
+		startTime(c, "total-request");
+
 		try {
+			startTime(c, "auth-check");
 			const { valid, key } = await auth.api.verifyApiKey({
 				body: {
 					key: c.req.valid("query").apiKey,
 				},
 			});
+			endTime(c, "auth-check");
 
 			if (!valid || !key || !key.metadata?.workspaceId) {
+				setMetric(c, "auth-failed", 1);
 				return c.json({ error: "Unauthorized" }, 401);
 			}
 
@@ -71,11 +80,21 @@ const screenshots = new OpenAPIHono<{ Variables: Variables }>().openapi(
 			const workspaceId = key.metadata.workspaceId;
 			const userId = key.userId;
 
-			// Ensure quota is available before continuing
+			const transformedParams = {
+				...queryParams,
+				selector: queryParams.selector?.trim() || undefined,
+			};
+
+			const cacheKey = generateCacheKey(workspaceId, transformedParams);
+			setMetric(c, "cache-key-generated", 1);
+
+			startTime(c, "quota-check");
 			try {
 				await assertQuotaAvailable(userId);
 			} catch (error) {
+				endTime(c, "quota-check");
 				if (error instanceof RequestQuotaError) {
+					setMetric(c, "quota-exceeded", 1);
 					return c.json(
 						{
 							error:
@@ -88,109 +107,114 @@ const screenshots = new OpenAPIHono<{ Variables: Variables }>().openapi(
 				}
 				throw error;
 			}
+			endTime(c, "quota-check");
 
-			let existingKey: string | null = null;
-			try {
-				existingKey = await getExistingScreenshotKey(workspaceId, queryParams);
-			} catch (error) {
-				logger.error({ err: error }, "error checking for existing screenshot");
+			const result = await deduplicateRequest(
+				cacheKey,
+				async () => {
+					startTime(c, "existing-check");
+					const existingKey = await getExistingScreenshotKey(
+						workspaceId,
+						transformedParams,
+					);
+					endTime(c, "existing-check");
 
-				existingKey = null;
-			}
+					if (existingKey) {
+						setMetric(c, "screenshot-found-existing", 1);
+						return {
+							key: existingKey,
+							created: false,
+						};
+					}
 
-			let raceResult: { key: string; created: boolean } | "timeout";
+					startTime(c, "job-enqueue");
+					const jobResult = await enqueueScreenshotJob(
+						workspaceId,
+						userId,
+						transformedParams,
+					);
+					endTime(c, "job-enqueue");
 
-			if (existingKey) {
-				raceResult = { key: existingKey, created: false };
-			} else {
-				// Not cached; enqueue generation and apply 15s timeout
-				const jobPromise = await enqueueScreenshotJob(
-					workspaceId,
-					userId,
-					queryParams,
-				);
+					setMetric(c, "screenshot-job-enqueued", 1);
+					logger.info(
+						{
+							workspaceId,
+							url: transformedParams.url,
+							cacheKey,
+						},
+						"Screenshot job enqueued successfully",
+					);
 
-				const TIMEOUT_MS = 15_000;
-				let timeoutId: ReturnType<typeof setTimeout> | null = null;
-				const timer = new Promise<"timeout">((resolve) => {
-					timeoutId = setTimeout(() => resolve("timeout" as const), TIMEOUT_MS);
-				});
-
-				raceResult = await Promise.race([jobPromise, timer]);
-
-				if (timeoutId) {
-					clearTimeout(timeoutId);
-				}
-			}
+					return {
+						key: jobResult.key || null,
+						created: jobResult.created,
+					};
+				},
+				c,
+			);
 
 			const headers = new Headers();
-			let body: ArrayBuffer | Buffer | null = null;
+			let body: ArrayBuffer | Buffer;
 			let contentType: string;
-			let createdFlag = false;
 
-			if (raceResult === "timeout") {
-				// Return 1x1 placeholder and allow client to retry
+			if (result.data.key) {
+				startTime(c, "s3-fetch");
+				try {
+					body = await s3.file(result.data.key).arrayBuffer();
+					endTime(c, "s3-fetch");
+					contentType = `image/${queryParams.format}`;
+
+					if (queryParams.isCached) {
+						const cacheTtl = queryParams.cacheTtl || 3600;
+						headers.set(
+							"Cache-Control",
+							`public, max-age=${cacheTtl}, s-maxage=${cacheTtl}, stale-while-revalidate=${cacheTtl}`,
+						);
+					}
+
+					setMetric(c, "screenshot-served", 1);
+				} catch (error) {
+					endTime(c, "s3-fetch");
+					logger.error(
+						{ err: error, key: result.data.key },
+						"Failed to fetch screenshot from S3",
+					);
+					body = PLACEHOLDER_IMAGE;
+					contentType = "image/png";
+					headers.set("Retry-After", "5");
+					setMetric(c, "screenshot-s3-error", 1);
+				}
+			} else {
 				body = PLACEHOLDER_IMAGE;
 				contentType = "image/png";
-				// Prevent caching of placeholder so user can retry quickly
+
 				headers.set(
 					"Cache-Control",
-					"no-cache, no-store, must-revalidate, proxy-revalidate, max-age=0",
+					"no-cache, no-store, must-revalidate, max-age=0",
 				);
-				headers.set("CDN-Cache-Control", "no-cache, no-store, max-age=0");
-			} else {
-				const { key: objectKey } = raceResult as {
-					key: string;
-					created: boolean;
-				};
-				createdFlag = (raceResult as { created: boolean }).created;
-				if (!objectKey) {
-					return c.json({ error: "Unauthorized" }, 401);
-				}
-				try {
-					body = await s3.file(objectKey).arrayBuffer();
-				} catch (e) {
-					return c.json({ error: "Screenshot not ready" }, 503, {
-						headers: ["Retry-After: 5"],
-					});
-				}
+				headers.set("Retry-After", "10");
 
-				contentType = `image/${queryParams.format}`;
-
-				if (queryParams.isCached) {
-					const cacheTtl = queryParams.cacheTtl;
-					headers.set(
-						"Cache-Control",
-						`public, max-age=${cacheTtl}, s-maxage=${cacheTtl}, stale-while-revalidate=${cacheTtl}, durable`,
-					);
-					headers.set(
-						"CDN-Cache-Control",
-						`max-age=${cacheTtl}, stale-while-revalidate=${cacheTtl}, durable`,
-					);
-				} else {
-					headers.set(
-						"Cache-Control",
-						"no-cache, no-store, must-revalidate, proxy-revalidate, max-age=0",
-					);
-					headers.set(
-						"CDN-Cache-Control",
-						"no-cache, no-store, must-revalidate, proxy-revalidate, max-age=0",
-					);
-				}
+				setMetric(c, "placeholder-served", 1);
 			}
 
-			if (createdFlag) {
-				const quota = await consumeQuota(userId, {
-					workspaceId,
-					url: queryParams.url,
-					format: queryParams.format,
-					userAgent: queryParams.userAgent,
-					source: "rest",
-				});
-				headers.set("X-Remaining-Requests", String(quota.remaining));
-				if (quota.nextRefillAt) {
-					headers.set("X-Refill-At", String(quota.nextRefillAt.getTime()));
+			if (result.data.created && !result.wasDeduplicated) {
+				startTime(c, "quota-consume");
+				try {
+					const quota = await consumeQuota(userId, {
+						workspaceId,
+						url: queryParams.url,
+						format: queryParams.format,
+						userAgent: queryParams.userAgent,
+						source: "rest-optimized",
+					});
+					headers.set("X-Remaining-Requests", String(quota.remaining));
+					if (quota.nextRefillAt) {
+						headers.set("X-Refill-At", String(quota.nextRefillAt.getTime()));
+					}
+				} catch (error) {
+					logger.error({ err: error }, "Failed to consume quota");
 				}
+				endTime(c, "quota-consume");
 			}
 
 			headers.set("Content-Length", String(body.byteLength));
@@ -201,13 +225,19 @@ const screenshots = new OpenAPIHono<{ Variables: Variables }>().openapi(
 				`inline; filename="screenshot.${queryParams.format}"`,
 			);
 
-			return c.body(body ?? new Uint8Array(), { headers });
-		} catch (error) {
-			const errorResponse = createErrorResponse(error, c.get("requestId"));
+			headers.set("X-Deduplicated", result.wasDeduplicated ? "true" : "false");
+			headers.set("X-Cache-Key", cacheKey);
 
-			return c.json(errorResponse, 400);
+			endTime(c, "total-request");
+			return c.body(body, { headers });
+		} catch (error) {
+			endTime(c, "total-request");
+			setMetric(c, "request-error", 1);
+
+			const errorResponse = createErrorResponse(error, c.get("requestId"));
+			return c.json(errorResponse, 500);
 		}
 	},
 );
 
-export default screenshots;
+export default optimizedScreenshots;
